Conflicts:
	docs/relintercalar/relatóriointercalar3183132466V1.docx

+++ test/Goertzel/goertzel.h
@@ -8,7 +8,7 @@
 #define PI (4*atan(1.0))  
 typedef short U16;
 
-float pot_freq_(U16 * x, int n, int fs,int fo);
+double pot_freq_(U16 * x, int n, int fs,int fo);
 
 float pot_freq(U16 * x, int n, int k);
